+++ test/navigation_scheme_test.dart
@@ -1,3 +1,5 @@
+// ignore_for_file: invalid_use_of_protected_member
+
 import 'package:flutter_test/flutter_test.dart';
 
 import 'package:theseus_navigator/theseus_navigator.dart';
@@ -84,23 +86,20 @@ void main() {
         navigationScheme = NavigationScheme(
           destinations: [
             TestDestinations.home,
-            TestDestinations.catalog,
-            TestDestinations.about,
+            TestDestinations.aboutWithInvalidRedirection,
             TestDestinations.login,
           ],
         );
       });
-      test('Current destination is stored after redirection', () async {
-        await navigationScheme.goTo(TestDestinations.login,
-            isRedirection: true);
+      test('Original destination is saved in the configuration of redirection destination', () async {
+        await navigationScheme.goTo(TestDestinations.aboutWithInvalidRedirection);
         expect(navigationScheme.currentDestination, TestDestinations.login);
-        expect(navigationScheme.redirectedFrom, TestDestinations.home);
+        expect(navigationScheme.currentDestination.configuration.redirectedFrom, TestDestinations.aboutWithInvalidRedirection);
       });
       test('User can navigate back from the redirected destination', () async {
-        await navigationScheme.goTo(TestDestinations.login,
-            isRedirection: true);
+        await navigationScheme.goTo(TestDestinations.aboutWithInvalidRedirection);
         navigationScheme.goBack();
-        expect(navigationScheme.currentDestination, TestDestinations.home);
+        expect(navigationScheme.currentDestination, TestDestinations.aboutWithInvalidRedirection);
       });
     });
     group('Error handling', () {
@@ -143,8 +142,8 @@ void main() {
       test(
           'For custom root navigator, if the error destination is provided, it should be included to the navigation scheme',
           () {
-        expect(
-            navigationScheme.findDestination('/error'), TestDestinations.error);
+        expect(navigationSchemeCustom.findDestination('/error'),
+            TestDestinations.error);
       });
       test(
           'Redirect to error destination when navigate to non-existent destination',